Refactoring TlsOkHttpClientBuilder by eliminating an indirection layer to construct OkHttpClient
